requirements.txt
Django
psycopg2
djangorestframework
django-cors-headers
pillow
djangorestframework-simplejwt